data/requirements.txt
yfinance
psycopg2-binary